code/RAGBench/requirements.txt
pytest
weaviate-client
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-weaviate==1.0.2
# nltk==3.8.1